requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & Migrations
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security (Token Validation Only)
python-jose[cryptography]==3.3.0

# Validation & Settings
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Utilities & Microservice Communication
redis==5.0.1
httpx==0.25.2

# gRPC & Protocol Buffers
grpcio==1.60.0
protobuf==4.25.1
grpcio-tools==1.60.0

# Testing
pytest==7.4.3


